+++ tg-adapter/src/main/java/com/iyuriy/notification/command/CommandContainer.java
@@ -2,7 +2,6 @@ package com.iyuriy.notification.command;
 
 import com.google.common.collect.ImmutableMap;
 import com.iyuriy.notification.repositories.UserRepository;
-import com.iyuriy.notification.services.RestEventSender;
 import org.springframework.stereotype.Service;
 
 import static com.iyuriy.notification.common.parser.UserEventType.*;
@@ -13,14 +12,12 @@ public class CommandContainer {
     private final ImmutableMap<String, Command> commandMap;
     private final Command unknownCommand;
 
-    public CommandContainer(RestEventSender restEventSender, UserRepository userRepository
-    ) {
+    public CommandContainer(UserRepository userRepository) {
 
         commandMap = ImmutableMap.<String, Command>builder()
-                .put(START.getUserEventType(), new StartCommand(userRepository))
-                .put(STOP.getUserEventType(), new StopCommand())
+                .put(START.getUserEventType(), new StartCommand())
+                .put(STOP.getUserEventType(), new StopCommand(userRepository))
                 .put(HELP.getUserEventType(), new HelpCommand())
-                .put(NO.getUserEventType(), new NoCommand())
                 .put(ADD.getUserEventType(), new AddCommand())
                 .put(TIME_ZONE.getUserEventType(), new TimeZoneCommand(userRepository))
                 .build();
