--- tg-adapter/src/main/java/com/iyuriy/notification/command/NoCommand.java
@@ -1,18 +1,0 @@
-package com.iyuriy.notification.command;
-
-import org.telegram.telegrambots.meta.api.objects.Update;
-
-public class NoCommand implements Command {
-
-
-    public static final String NO_MESSAGE =
-            "Поддерживаются команды,\n"
-                    + "начинающиеся со слэша(/).\n"
-                    + "Чтобы посмотреть список\n"
-                    + "доступных команд введите /help";
-
-    @Override
-    public String execute(Update update) {
-        return NO_MESSAGE;
-    }
-}
