+++ tg-adapter/src/main/java/com/iyuriy/notification/repositories/UserRepository.java
@@ -8,4 +8,6 @@ import org.springframework.stereotype.Repository;
 public interface UserRepository extends JpaRepository<User, Long> {
 
     User findByChatId (Long chatId);
+
+    User deleteByChatId (Long chatId);
 }
